Key: Refactored to store KeySym instead of KeyCode.
The Keycode might point to a different key depending on the keymap.
When started from .xinitrc, the keymap may still be set to "us", so
if a user binds W-x, the keycode for "x" might become the keycode for
"z" after switching to a qwertz layout.

Using the keysymbol and getting the keycode from X each time we need it
solves this. That's what symbols are for, after all.