Fixed "wrong keymap during .xinitrc" bug, updated ChangeLog and manual.
What a nasty bug. It turns out that the X server won't actually load
a user specified keymap until the first key is pressed. This would
cause a prefix key like "W-y" to be grabbed with the wrong keycode,
e.g (29 =~ "z").

The solution is simple: regrab the keys every time after processing
a key event in mainloop. The first modifier pressed will pass through
and cause the X server to load the proper keymap, upon which the keys
are regrabbed to the proper keycodes... thus the first "W-y" will work
in any case.